refactor: split events and anchors on it's own file
